+++ src/js/components/parallax.js
@@ -295,13 +295,13 @@
         url     = element.css('background-image').replace(/^url\(/g, '').replace(/\)$/g, '').replace(/("|')/g, '');
         check   = function() {
 
-            var w = element.width(), h = element.height(), extra = (prop=='bg') ? opts.diff : (opts.diff/100) * h;
+            var w = element.innerWidth(), h = element.innerHeight(), extra = (prop=='bg') ? opts.diff : (opts.diff/100) * h;
 
             h += extra;
             w += Math.ceil(extra * ratio);
 
             if (w-extra < size.w && h < size.h) {
-                return obj.element.css({'background-size': ''});
+                return obj.element.css({'background-size': 'auto'});
             }
 
             // if element height < parent height (gap underneath)