+++ src/js/components/sticky.js
@@ -301,8 +301,7 @@
                     sticky.element.css({
                         position : "fixed",
                         top      : newTop,
-                        width    : sticky.getWidthFrom.length ? sticky.getWidthFrom.width() : sticky.element.width(),
-                        left     : sticky.wrapper.offset().left
+                        width    : sticky.getWidthFrom.length ? sticky.getWidthFrom.width() : sticky.element.width()
                     });
 
                     if (!sticky.init) {