+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 # WIP
 
   - Fixed Animated slideset filter
+  - Fixed noDragClass option (sortable)
 
 ### 2.24.3 (December 18, 2015)