+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 
 # WIP
+  - Added grid parallax component
   - Added new grid-width classes to grid component
   - Added better vertical-align for embedded content to base component
   - Added :before and :after to print component
-  - Added gridparallax component
   - Fixed Animated slideset filter
   - Fixed noDragClass option (sortable)
   - Fixed drop target detection (sortable)